+++ example/android/app/src/main/java/com/fidenz/mapxus_positioning_flutter_example/MainActivity.java
@@ -1,6 +1,0 @@
-package com.fidenz.mapxus_positioning_flutter_example;
-
-import io.flutter.embedding.android.FlutterActivity;
-
-public class MainActivity extends FlutterActivity {
-}