+++ example/android/app/src/main/kotlin/com/fidenz/mapxus_positioning_flutter_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.fidenz.mapxus_positioning_flutter_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()